+++ src/components/LanguagesDropdown/index.tsx
@@ -2,7 +2,9 @@
 import { usePathname } from 'next/navigation';
 import { Globe } from 'react-feather';
 import DropdownIcon, { DropdownItem } from '@components/DropdownIcon';
+import { useTranslator } from '@hooks/useTanslator';
 import { locales } from '@src/i18n';
+import { Suspense } from 'react';
 
 
 function trimLocaleFrom(pathname: string) {
@@ -13,14 +15,23 @@ function trimLocaleFrom(pathname: string) {
     return pathname;
 }
 
-export default function LanguagesDropdown() {
+export async function LanguagesDropdown() {
     const pathname = usePathname();
+    const t = await useTranslator();
     const trimmedPathname = trimLocaleFrom(pathname);
 
     return (
         <DropdownIcon icon={Globe}>
-            <DropdownItem href={`/es${trimmedPathname}`}>Español</DropdownItem>
-            <DropdownItem href={`/en${trimmedPathname}`}>Ingles</DropdownItem>
+            <DropdownItem href={`/es${trimmedPathname}`}>{t('languagesDropdown.es')}</DropdownItem>
+            <DropdownItem href={`/en${trimmedPathname}`}>{t('languagesDropdown.en')}</DropdownItem>
         </DropdownIcon>
+    );
+}
+
+export default function LazyLanguagesDropdown() {
+    return (
+        <Suspense fallback={<DropdownIcon icon={Globe} />}>
+            <LanguagesDropdown />
+        </Suspense>
     );
 }
