+++ src/dictionaries/en.json
@@ -1,2 +1,6 @@
 {
+    "languagesDropdown": {
+        "es": "Spanish",
+        "en": "English"
+    }
 }
