+++ src/dictionaries/es.json
@@ -1,2 +1,6 @@
 {
+    "languagesDropdown": {
+        "es": "Español",
+        "en": "Inglés"
+    }
 }
